+++ client/go.mod
@@ -3,7 +3,7 @@ module github.com/TUM-Dev/Campus-Backend/client
 go 1.18
 
 require (
-	github.com/TUM-Dev/Campus-Backend/api v0.0.0-20220921171658-c0ee084ed389
+	github.com/TUM-Dev/Campus-Backend/server/api
 	github.com/sirupsen/logrus v1.9.0
 	google.golang.org/grpc v1.49.0
 	google.golang.org/protobuf v1.28.1
