+++ server/api/gen/openapiv2/CampusService.swagger.json
