+++ server/api/generate.sh
