+++ server/api/installBuf.sh
